fix: 🐛 fix more risky linter warnings
