+++ src/json-cli/test/spec.json.ts
@@ -168,7 +168,7 @@ const testCases: TestCase[] = [
     },
     patch: [
       // @ts-ignore
-      {op: 'add', path: '/baz', value: 'qux', op: 'remove'},
+      {path: '/baz', value: 'qux', op: 'remove'},
     ],
     error: "operation has two 'op' members",
     disabled: true,
