+++ src/json-crdt-peritext-ui/react/selection/Caret.tsx
@@ -1,4 +1,4 @@
-import type * as React from 'react';
+import * as React from 'react';
 import {Char} from '../../constants';
 import {usePeritext} from '../context';
 import {useCaret} from './hooks';
